feat(cli): split 'automations prompt' into 'prompt get' and 'prompt set'
The mixed read/write `automations prompt <id>` command was a footgun — without
`--from-file` it tried to read from stdin and bailed on EOF, so stdout
redirects produced empty files and there was no scriptable read path. Split
into explicit subcommands:

- `automations prompt get <id>` writes the prompt to stdout (no trailing
  newline added) so `get | diff -` and `get > file.md` round-trip cleanly.
- `automations prompt set <id> --from-file <path>` replaces the prompt;
  `--from-file -` reads from stdin.
