+++ packages/cli/src/commands/automations/get/command.ts
@@ -7,8 +7,9 @@ export default command({
 	run: async ({ ctx, args }) => {
 		const id = args.id as string;
 		const result = await ctx.api.automation.get.query({ id });
-		// Prompt is fetched via `superset automations prompt <id>` (it can be
-		// large markdown). Runs are paginated via `superset automations logs <id>`.
+		// Prompt is fetched via `superset automations prompt get <id>` (it can
+		// be large markdown). Runs are paginated via `superset automations
+		// logs <id>`.
 		const { prompt: _prompt, ...automation } = result;
 		return { data: automation };
 	},
