+++ packages/cli/src/commands/automations/prompt/meta.ts
@@ -1,0 +1,3 @@
+export default {
+	description: "Read or write an automation's prompt",
+};
